Moved mocks to cli_tools/mocks due to a circular dependency between cli_tools and mocks and the fact mocks are only used in cli_tools
