+++ cli_tools/gce_ovf_import/ovf_importer/ovf_importer_test.go
@@ -31,8 +31,8 @@ import (
 
 	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/common/utils/logging"
 	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/gce_ovf_import/ovf_import_params"
+	"github.com/GoogleCloudPlatform/compute-image-tools/cli_tools/mocks"
 	"github.com/GoogleCloudPlatform/compute-image-tools/daisy"
-	"github.com/GoogleCloudPlatform/compute-image-tools/mocks"
 )
 
 const instanceImportWorkflowPath = "../../test_data/test_import_ovf_to_instance.wf.json"
