+++ cli_tools/mocks/mock_compute_client.go
