+++ cli_tools/mocks/mock_compute_service.go
