+++ cli_tools/mocks/mock_metadata_gce.go
