+++ cli_tools/mocks/mock_object_deleter.go
