+++ cli_tools/mocks/mock_object_iterator_creator.go
