+++ cli_tools/mocks/mock_ovf_descriptor_loader.go
