+++ cli_tools/mocks/mock_resource_location_retriever.go
