+++ cli_tools/mocks/mock_scratch_bucket_creator.go
