+++ cli_tools/mocks/mock_storage_client.go
