--- mocks/go.mod
@@ -1,18 +1,0 @@
-module github.com/GoogleCloudPlatform/compute-image-tools/mocks
-
-go 1.13
-
-require (
-	cloud.google.com/go v0.56.0 // indirect
-	cloud.google.com/go/storage v1.6.0
-	github.com/GoogleCloudPlatform/compute-image-tools/cli_tools v0.0.0-20200406182414-bf9021434372
-	github.com/GoogleCloudPlatform/compute-image-tools/daisy v0.0.0-20200406182414-bf9021434372
-	github.com/golang/mock v1.4.3
-	github.com/vmware/govmomi v0.22.2
-	golang.org/x/exp v0.0.0-20200228211341-fcea875c7e85 // indirect
-	golang.org/x/net v0.0.0-20200324143707-d3edc9973b7e // indirect
-	golang.org/x/sys v0.0.0-20200406155108-e3b113bbe6a4 // indirect
-	golang.org/x/tools v0.0.0-20200406172401-903869a8272d // indirect
-	google.golang.org/api v0.21.0
-	google.golang.org/genproto v0.0.0-20200406120821-33397c535dc2 // indirect
-)
